refactor(workspace): promote the site dialog to a Workspace Settings modal
The site.json editor was buried under File > Export > Website Settings
and read as a one-off. Workspace-scoped configuration now has its own
home: File > Workspace Settings (next to the global Settings item on
Windows/Linux, in the File menu on macOS), a tabbed modal like the
global one with Website as the first tab, still saved to .glyph/ so it
is shared with the workspace. New workspaceSettings i18n namespace and
menu/commands keys in all five locales; menu id, event, palette
command, and CSS scope renamed to match.
