+++ README.md
@@ -72,7 +72,7 @@ The [`samples/`](samples) directory is a tiny demo workspace. Open it as a folde
 - Resizable panels: drag the edge of the files sidebar, outline sidebar, AI chat panel, or the backlinks divider to resize it (double-click resets); sizes persist across restarts
 - Print: `Cmd/Ctrl+P` with configurable page breaks, optional TOC, and theme-color control
 - Export to HTML, Word (DOCX), EPUB, and PDF via `File → Export`: works for markdown documents and Jupyter notebooks, writes a file directly (no print dialog), and reuses the rendered output (math, code highlighting, tables, images inlined) so files are self-contained and offline. Mermaid and D2 diagrams and SVG images embed in the PDF as true vectors, so they stay crisp at any zoom. Task-list checkboxes are read-only in exports; exported HTML follows the reader's light/dark system preference.
-- Export an entire workspace as a static website via `File → Export → Website…` (folder workspaces): one linked HTML page per markdown file with shared styles, cross-page navigation, a per-page outline, an index built from the root index.md (falling back to the root README, then a generated page list), resolved wikilinks and relative links, copied images, and inline Mermaid diagrams. File > Export > Website Settings… edits it in-app; a `.glyph/site.json` at the workspace root optionally sets the site title, description, favicon, social-preview (Open Graph) tags, robots.txt, and the theme: a polished GitHub-style look by default (with a site header on every page linking home), "plain" for the minimal one, and plugins can contribute their own; every page gets a "Page · Site" title either way. Also available headless from the CLI (`glyph <folder> --export-website <outDir>`) for CI publishing, e.g. to GitHub Pages
+- Export an entire workspace as a static website via `File → Export → Website…` (folder workspaces): one linked HTML page per markdown file with shared styles, cross-page navigation, a per-page outline, an index built from the root index.md (falling back to the root README, then a generated page list), resolved wikilinks and relative links, copied images, and inline Mermaid diagrams. File > Workspace Settings… edits it in-app (Website tab); a `.glyph/site.json` at the workspace root optionally sets the site title, description, favicon, social-preview (Open Graph) tags, robots.txt, and the theme: a polished GitHub-style look by default (with a site header on every page linking home), "plain" for the minimal one, and plugins can contribute their own; every page gets a "Page · Site" title either way. Also available headless from the CLI (`glyph <folder> --export-website <outDir>`) for CI publishing, e.g. to GitHub Pages
 - Live reload: file watcher auto-updates on external changes
 - Undo / redo for in-document edits: `Cmd/Ctrl+Z` and `Cmd/Ctrl+Shift+Z` reverse task-list checkbox toggles and other programmatic edits per tab
 - Drag and drop markdown files, notebooks, canvases, or folders to open
