+++ src/hooks/useMenuEvents.test.ts
@@ -23,7 +23,7 @@ function noopHandlers(overrides: Partial<MenuEventHandlers> = {}): MenuEventHand
     exportEpub: vi.fn(),
     exportPdf: vi.fn(),
     exportWebsite: vi.fn(),
-    websiteSettings: vi.fn(),
+    workspaceSettings: vi.fn(),
     zoomIn: vi.fn(),
     zoomOut: vi.fn(),
     zoomReset: vi.fn(),
@@ -66,7 +66,7 @@ describe("useMenuEvents", () => {
         "menu-export-epub",
         "menu-export-pdf",
         "menu-export-website",
-        "menu-export-website-settings",
+        "menu-workspace-settings",
         "menu-zoom-in",
         "menu-zoom-out",
         "menu-zoom-reset",
