+++ src/hooks/useMenuEvents.ts
@@ -20,7 +20,7 @@ export interface MenuEventHandlers {
   exportEpub: () => void;
   exportPdf: () => void;
   exportWebsite: () => void;
-  websiteSettings: () => void;
+  workspaceSettings: () => void;
   zoomIn: () => void;
   zoomOut: () => void;
   zoomReset: () => void;
@@ -56,7 +56,7 @@ export function useMenuEvents(handlers: MenuEventHandlers) {
       subscribe("menu-export-epub", handlers.exportEpub),
       subscribe("menu-export-pdf", handlers.exportPdf),
       subscribe("menu-export-website", handlers.exportWebsite),
-      subscribe("menu-export-website-settings", handlers.websiteSettings),
+      subscribe("menu-workspace-settings", handlers.workspaceSettings),
       subscribe("menu-zoom-in", handlers.zoomIn),
       subscribe("menu-zoom-out", handlers.zoomOut),
       subscribe("menu-zoom-reset", handlers.zoomReset),
