+++ src/locales/de/commands.json
@@ -27,7 +27,7 @@
   "exportEpub": "Als EPUB exportieren…",
   "exportPdf": "Als PDF exportieren…",
   "exportWebsite": "Arbeitsbereich als Website exportieren…",
-  "websiteSettings": "Website-Einstellungen…",
+  "workspaceSettings": "Arbeitsbereich-Einstellungen…",
   "zoomIn": "Vergrößern",
   "zoomOut": "Verkleinern",
   "zoomReset": "Zoom zurücksetzen",
