+++ src/locales/de/menu.json
@@ -14,7 +14,7 @@
   "exportEpub": "EPUB…",
   "exportPdf": "PDF…",
   "exportWebsite": "Website…",
-  "exportWebsiteSettings": "Website-Einstellungen…",
+  "workspaceSettings": "Arbeitsbereich-Einstellungen…",
   "closeTab": "Tab schließen",
   "close": "Fenster schließen",
   "settings": "Einstellungen…",
