+++ src/locales/en/commands.json
@@ -27,7 +27,7 @@
   "exportEpub": "Export to EPUB…",
   "exportPdf": "Export to PDF…",
   "exportWebsite": "Export Workspace as Website…",
-  "websiteSettings": "Website Settings…",
+  "workspaceSettings": "Workspace Settings…",
   "zoomIn": "Zoom In",
   "zoomOut": "Zoom Out",
   "zoomReset": "Reset Zoom",
