+++ src/locales/en/menu.json
@@ -14,7 +14,7 @@
   "exportEpub": "EPUB…",
   "exportPdf": "PDF…",
   "exportWebsite": "Website…",
-  "exportWebsiteSettings": "Website Settings…",
+  "workspaceSettings": "Workspace Settings…",
   "closeTab": "Close Tab",
   "close": "Close Window",
   "settings": "Settings…",
