+++ src/locales/es/commands.json
@@ -27,7 +27,7 @@
   "exportEpub": "Exportar a EPUB…",
   "exportPdf": "Exportar a PDF…",
   "exportWebsite": "Exportar espacio de trabajo como sitio web…",
-  "websiteSettings": "Configuración del sitio web…",
+  "workspaceSettings": "Configuración del espacio de trabajo…",
   "zoomIn": "Acercar",
   "zoomOut": "Alejar",
   "zoomReset": "Restablecer zoom",
