+++ src/locales/es/menu.json
@@ -14,7 +14,7 @@
   "exportEpub": "EPUB…",
   "exportPdf": "PDF…",
   "exportWebsite": "Sitio web…",
-  "exportWebsiteSettings": "Configuración del sitio web…",
+  "workspaceSettings": "Configuración del espacio de trabajo…",
   "closeTab": "Cerrar pestaña",
   "close": "Cerrar ventana",
   "settings": "Ajustes…",
