+++ src/locales/fa/commands.json
@@ -27,7 +27,7 @@
   "exportEpub": "خروجی به EPUB…",
   "exportPdf": "خروجی به PDF…",
   "exportWebsite": "خروجی فضای کاری به‌صورت وب‌سایت…",
-  "websiteSettings": "تنظیمات وب‌سایت…",
+  "workspaceSettings": "تنظیمات فضای کاری…",
   "zoomIn": "بزرگ‌نمایی",
   "zoomOut": "کوچک‌نمایی",
   "zoomReset": "بازنشانی بزرگ‌نمایی",
