+++ src/locales/fa/menu.json
@@ -14,7 +14,7 @@
   "exportEpub": "EPUB…",
   "exportPdf": "PDF…",
   "exportWebsite": "وب‌سایت…",
-  "exportWebsiteSettings": "تنظیمات وب‌سایت…",
+  "workspaceSettings": "تنظیمات فضای کاری…",
   "closeTab": "بستن زبانه",
   "close": "بستن پنجره",
   "settings": "تنظیمات…",
