+++ src/locales/fa/workspaceSettings.json
@@ -1,0 +1,45 @@
+{
+  "modal": {
+    "heading": "تنظیمات فضای کاری",
+    "close": "بستن"
+  },
+  "tabs": {
+    "website": "وب‌سایت"
+  },
+  "empty": "برای ویرایش تنظیمات، یک فضای کاری پوشه باز کنید.",
+  "website": {
+    "description": "نحوه خروجی این فضای کاری به‌صورت وب‌سایت. در .glyph/site.json ذخیره می‌شود و همراه پوشه می‌ماند.",
+    "fields": {
+      "title": {
+        "label": "عنوان سایت",
+        "hint": "زبانه مرورگر، سرصفحه و کارت‌های اجتماعی"
+      },
+      "description": {
+        "label": "توضیحات"
+      },
+      "baseUrl": {
+        "label": "نشانی پایه",
+        "hint": "پیوندها و تصاویر اجتماعی را فعال می‌کند"
+      },
+      "favicon": {
+        "label": "مسیر فاوآیکون",
+        "hint": "نسبت به فضای کاری؛ favicon.ico/png/svg در ریشه خودکار شناسایی می‌شود"
+      },
+      "socialImage": {
+        "label": "مسیر تصویر اجتماعی",
+        "hint": "در پیش‌نمایش پیوندها نمایش داده می‌شود؛ نیازمند نشانی پایه"
+      },
+      "robots": {
+        "label": "موتورهای جستجو (robots.txt)",
+        "none": "robots.txt نوشته نشود",
+        "all": "اجازه به همه",
+        "disallow": "مسدود کردن همه"
+      },
+      "theme": {
+        "label": "پوسته"
+      }
+    },
+    "save": "ذخیره",
+    "cancel": "لغو"
+  }
+}
