+++ src/locales/zh/commands.json
@@ -27,7 +27,7 @@
   "exportEpub": "导出为 EPUB…",
   "exportPdf": "导出为 PDF…",
   "exportWebsite": "将工作区导出为网站…",
-  "websiteSettings": "网站设置…",
+  "workspaceSettings": "工作区设置…",
   "zoomIn": "放大",
   "zoomOut": "缩小",
   "zoomReset": "重置缩放",
