+++ src/locales/zh/menu.json
@@ -14,7 +14,7 @@
   "exportEpub": "EPUB…",
   "exportPdf": "PDF…",
   "exportWebsite": "网站…",
-  "exportWebsiteSettings": "网站设置…",
+  "workspaceSettings": "工作区设置…",
   "closeTab": "关闭标签页",
   "close": "关闭窗口",
   "settings": "设置…",
