+++ src/locales/zh/workspaceSettings.json
@@ -1,0 +1,45 @@
+{
+  "modal": {
+    "heading": "工作区设置",
+    "close": "关闭"
+  },
+  "tabs": {
+    "website": "网站"
+  },
+  "empty": "打开一个文件夹工作区以编辑其设置。",
+  "website": {
+    "description": "此工作区导出为网站的方式。保存到 .glyph/site.json，随文件夹一起迁移。",
+    "fields": {
+      "title": {
+        "label": "网站标题",
+        "hint": "浏览器标签、页眉和社交卡片"
+      },
+      "description": {
+        "label": "描述"
+      },
+      "baseUrl": {
+        "label": "基础 URL",
+        "hint": "启用社交链接和图片"
+      },
+      "favicon": {
+        "label": "网站图标路径",
+        "hint": "相对于工作区；根目录的 favicon.ico/png/svg 会自动识别"
+      },
+      "socialImage": {
+        "label": "社交图片路径",
+        "hint": "显示在链接预览中；需要基础 URL"
+      },
+      "robots": {
+        "label": "搜索引擎 (robots.txt)",
+        "none": "不写入 robots.txt",
+        "all": "全部允许",
+        "disallow": "全部禁止"
+      },
+      "theme": {
+        "label": "主题"
+      }
+    },
+    "save": "保存",
+    "cancel": "取消"
+  }
+}
