+++ src/styles/settings.css
@@ -384,35 +384,35 @@
 }
 
 /* Cloud Sync modal */
-:is(.settings-sync, .settings-site) .settings-field {
+:is(.settings-sync, .settings-workspace) .settings-field {
   display: flex;
   flex-direction: column;
   gap: 4px;
   margin-bottom: 14px;
 }
-:is(.settings-sync, .settings-site) .settings-field-label {
+:is(.settings-sync, .settings-workspace) .settings-field-label {
   font-size: 13px;
   font-weight: 500;
   color: var(--color-text-primary);
 }
-:is(.settings-sync, .settings-site) .settings-field-hint {
+:is(.settings-sync, .settings-workspace) .settings-field-hint {
   font-size: 11px;
   color: var(--color-text-secondary);
   font-weight: normal;
 }
-:is(.settings-sync, .settings-site) .settings-section-description {
+:is(.settings-sync, .settings-workspace) .settings-section-description {
   margin: 0 0 16px;
   font-size: 12px;
   color: var(--color-text-secondary);
 }
-:is(.settings-sync, .settings-site) .settings-empty,
-:is(.settings-sync, .settings-site) .settings-busy {
+:is(.settings-sync, .settings-workspace) .settings-empty,
+:is(.settings-sync, .settings-workspace) .settings-busy {
   padding: 24px 0;
   text-align: center;
   color: var(--color-text-secondary);
   font-size: 13px;
 }
-:is(.settings-sync, .settings-site) .settings-error {
+:is(.settings-sync, .settings-workspace) .settings-error {
   margin: 8px 0 0;
   padding: 8px 10px;
   font-size: 12px;
@@ -420,7 +420,7 @@
   background: rgba(229, 62, 62, 0.08);
   border-radius: var(--glyph-radius-sm);
 }
-:is(.settings-sync, .settings-site) .settings-actions {
+:is(.settings-sync, .settings-workspace) .settings-actions {
   display: flex;
   flex-wrap: wrap;
   gap: 8px;
